Implement cookie sessions as a generic type
Implements the concept of cookie storage and sessions by introducing
a generic type on `Client`, `ClientBuilder`, and related structs.
Ends up percolating through most types, and introduces some poor
ergonomics.

Necessitates the introduction of a `NullSession` no-op store, in order
to be able to specify the generic type in cases where no session is
wanted. Despite this, we also retain the `Option<S>` pattern for the
session, so we have the ability to `Client::take_session()`, leaving
no session in place on the client.